introduce Exec.set_code to fix vm.etch
