+++ src/halmos/sevm.py
@@ -8,7 +8,6 @@ from copy import deepcopy
 from collections import defaultdict
 from dataclasses import dataclass, field
 from functools import reduce
-from subprocess import Popen, PIPE
 from typing import (
     Any,
     Dict,
@@ -25,76 +24,19 @@ from z3 import *
 
 from .cheatcodes import halmos_cheat_code, hevm_cheat_code, console, Prank
 from .exceptions import *
-from .utils import (
-    assert_address,
-    assert_uint256,
-    bv_value_to_bytes,
-    byte_length,
-    color_info,
-    con_addr,
-    create_solver,
-    EVM,
-    hexify,
-    int_of,
-    restore_precomputed_hashes,
-    sha3_inv,
-    str_opcode,
-    unbox_int,
-)
+from .utils import *
 from .warnings import (
     warn,
     LIBRARY_PLACEHOLDER,
     INTERNAL_ERROR,
 )
 
-Word = Any  # z3 expression (including constants)
-Byte = Any  # z3 expression (including constants)
-Bytes = Any  # z3 expression (including constants)
-Address = BitVecRef  # 160-bitvector
 
 Steps = Dict[int, Dict[str, Any]]  # execution tree
 
 EMPTY_BYTES = b""
 MAX_CALL_DEPTH = 1024
 
-
-# dynamic BitVecSort sizes
-class BitVecSortCache:
-    def __init__(self):
-        self.cache = {}
-        for size in (
-            1,
-            8,
-            16,
-            32,
-            64,
-            128,
-            160,
-            256,
-            264,
-            288,
-            512,
-            544,
-            800,
-            1024,
-            1056,
-        ):
-            self.cache[size] = BitVecSort(size)
-
-    def __getitem__(self, size: int) -> BitVecSort:
-        hit = self.cache.get(size)
-        return hit if hit is not None else BitVecSort(size)
-
-
-BitVecSorts = BitVecSortCache()
-
-# known, fixed BitVecSort sizes
-BitVecSort1 = BitVecSorts[1]
-BitVecSort8 = BitVecSorts[8]
-BitVecSort160 = BitVecSorts[160]
-BitVecSort256 = BitVecSorts[256]
-BitVecSort264 = BitVecSorts[264]
-BitVecSort512 = BitVecSorts[512]
 
 # symbolic states
 # calldataload(index)
@@ -170,10 +112,6 @@ def id_str(x: Any) -> str:
     return hexify(x).replace(" ", "")
 
 
-def name_of(x: str) -> str:
-    return re.sub(r"\s+", "_", x)
-
-
 def padded_slice(lst: List, start: int, size: int, default=0) -> List:
     """
     Return a slice of lst, starting at start and with size elements. If the slice
@@ -222,38 +160,6 @@ def mnemonic(opcode) -> str:
         return str_opcode.get(opcode, hex(opcode))
     else:
         return str(opcode)
-
-
-def concat(args):
-    if len(args) > 1:
-        return Concat(args)
-    else:
-        return args[0]
-
-
-def uint256(x: BitVecRef) -> BitVecRef:
-    bitsize = x.size()
-    if bitsize > 256:
-        raise ValueError(x)
-    if bitsize == 256:
-        return x
-    return simplify(ZeroExt(256 - bitsize, x))
-
-
-def uint160(x: BitVecRef) -> BitVecRef:
-    bitsize = x.size()
-    if bitsize > 256:
-        raise ValueError(x)
-    if bitsize == 160:
-        return x
-    if bitsize > 160:
-        return simplify(Extract(159, 0, x))
-    else:
-        return simplify(ZeroExt(160 - bitsize, x))
-
-
-def con(n: int, size_bits=256) -> Word:
-    return BitVecVal(n, BitVecSorts[size_bits])
 
 
 def instruction_length(opcode: Any) -> int:
@@ -348,99 +254,6 @@ def wstore_bytes(
         if not eq(arr[i].sort(), BitVecSort8):
             raise ValueError(arr)
         mem[loc + i] = arr[i]
-
-
-def extract_bytes(data: BitVecRef, byte_offset: int, size_bytes: int) -> BitVecRef:
-    """Extract bytes from calldata. Zero-pad if out of bounds."""
-    n = data.size()
-    if n % 8 != 0:
-        raise ValueError(n)
-
-    # will extract hi - lo + 1 bits
-    hi = n - 1 - byte_offset * 8
-    lo = n - byte_offset * 8 - size_bytes * 8
-    lo = 0 if lo < 0 else lo
-
-    val = simplify(Extract(hi, lo, data))
-
-    zero_padding = size_bytes * 8 - val.size()
-    if zero_padding < 0:
-        raise ValueError(val)
-    if zero_padding > 0:
-        val = simplify(Concat(val, con(0, zero_padding)))
-
-    return val
-
-
-def extract_funsig(calldata: BitVecRef):
-    """Extracts the function signature (first 4 bytes) from calldata"""
-    return extract_bytes(calldata, 0, 4)
-
-
-def extract_string_argument(calldata: BitVecRef, arg_idx: int):
-    """Extracts idx-th argument of string from calldata"""
-    string_offset = int_of(
-        extract_bytes(calldata, 4 + arg_idx * 32, 32),
-        "symbolic offset for string argument",
-    )
-    string_length = int_of(
-        extract_bytes(calldata, 4 + string_offset, 32),
-        "symbolic size for string argument",
-    )
-    if string_length == 0:
-        return ""
-    string_value = int_of(
-        extract_bytes(calldata, 4 + string_offset + 32, string_length),
-        "symbolic string argument",
-    )
-    string_bytes = string_value.to_bytes(string_length, "big")
-    return string_bytes.decode("utf-8")
-
-
-def extract_string_array_argument(calldata: BitVecRef, arg_idx: int):
-    """Extracts idx-th argument of string array from calldata"""
-
-    array_slot = int_of(extract_bytes(calldata, 4 + 32 * arg_idx, 32))
-    num_strings = int_of(extract_bytes(calldata, 4 + array_slot, 32))
-
-    string_array = []
-
-    for i in range(num_strings):
-        string_offset = int_of(
-            extract_bytes(calldata, 4 + array_slot + 32 * (i + 1), 32)
-        )
-        string_length = int_of(
-            extract_bytes(calldata, 4 + array_slot + 32 + string_offset, 32)
-        )
-        string_value = int_of(
-            extract_bytes(
-                calldata, 4 + array_slot + 32 + string_offset + 32, string_length
-            )
-        )
-        string_bytes = string_value.to_bytes(string_length, "big")
-        string_array.append(string_bytes.decode("utf-8"))
-
-    return string_array
-
-
-def stringified_bytes_to_bytes(string_bytes: str):
-    """Converts a string of bytes to a bytes memory type"""
-
-    string_bytes_len = (len(string_bytes) + 1) // 2
-    string_bytes_len_enc = hex(string_bytes_len).replace("0x", "").rjust(64, "0")
-
-    string_bytes_len_ceil = (string_bytes_len + 31) // 32 * 32
-
-    ret_bytes = (
-        "00" * 31
-        + "20"
-        + string_bytes_len_enc
-        + string_bytes.ljust(string_bytes_len_ceil * 2, "0")
-    )
-    ret_len = len(ret_bytes) // 2
-    ret_bytes = bytes.fromhex(ret_bytes)
-
-    return BitVecVal(int.from_bytes(ret_bytes, "big"), ret_len * 8)
 
 
 @dataclass(frozen=True)
@@ -874,6 +687,15 @@ class Exec:  # an execution path
 
     def current_instruction(self) -> Instruction:
         return self.pgm.decode_instruction(self.pc)
+
+    def set_code(
+        self, who: Address, code: UnionType[bytes, BitVecRef, Contract]
+    ) -> None:
+        """
+        Sets the code at a given address.
+        """
+        assert_address(who)
+        self.code[who] = code if isinstance(code, Contract) else Contract(code)
 
     def str_cnts(self) -> str:
         return "".join(
@@ -1315,31 +1137,6 @@ class GenericStorage(Storage):
 
 
 SomeStorage = TypeVar("SomeStorage", bound=Storage)
-
-
-#             x  == b   if sort(x) = bool
-# int_to_bool(x) == b   if sort(x) = int
-def test(x: Word, b: bool) -> Word:
-    if is_bool(x):
-        if b:
-            return x
-        else:
-            return Not(x)
-    elif is_bv(x):
-        if b:
-            return x != con(0)
-        else:
-            return x == con(0)
-    else:
-        raise ValueError(x)
-
-
-def is_non_zero(x: Word) -> Word:
-    return test(x, True)
-
-
-def is_zero(x: Word) -> Word:
-    return test(x, False)
 
 
 def bitwise(op, x: Word, y: Word) -> Word:
@@ -1844,282 +1641,20 @@ class SEVM:
                 ex.solver.add(exit_code_var != con(0))
                 ret = arg
 
-            # TODO: factor out cheatcode semantics
             # halmos cheat code
             if eq(to, halmos_cheat_code.address):
                 ex.solver.add(exit_code_var != con(0))
-
-                funsig: int = int_of(
-                    extract_funsig(arg), "symbolic halmos cheatcode function selector"
-                )
-
-                if self.options.get("debug"):
-                    print(f"Executing halmos cheat code: {hex(funsig)}")
-
-                # createUint(uint256,string) returns (uint256)
-                if funsig == halmos_cheat_code.create_uint:
-                    bit_size = int_of(
-                        extract_bytes(arg, 4, 32),
-                        "symbolic bit size for halmos.createUint()",
-                    )
-                    name = name_of(extract_string_argument(arg, 1))
-                    if bit_size <= 256:
-                        label = f"halmos_{name}_uint{bit_size}_{ex.new_symbol_id():>02}"
-                        ret = uint256(BitVec(label, BitVecSorts[bit_size]))
-                    else:
-                        raise HalmosException(f"bitsize larger than 256: {bit_size}")
-
-                # createBytes(uint256,string) returns (bytes)
-                elif funsig == halmos_cheat_code.create_bytes:
-                    byte_size = int_of(
-                        extract_bytes(arg, 4, 32),
-                        "symbolic byte size for halmos.createBytes()",
-                    )
-                    name = name_of(extract_string_argument(arg, 1))
-                    label = f"halmos_{name}_bytes_{ex.new_symbol_id():>02}"
-                    symbolic_bytes = BitVec(label, BitVecSorts[byte_size * 8])
-                    ret = Concat(con(32), con(byte_size), symbolic_bytes)
-
-                # createUint256(string) returns (uint256)
-                elif funsig == halmos_cheat_code.create_uint256:
-                    name = name_of(extract_string_argument(arg, 0))
-                    label = f"halmos_{name}_uint256_{ex.new_symbol_id():>02}"
-                    ret = BitVec(label, BitVecSort256)
-
-                # createBytes32(string) returns (bytes32)
-                elif funsig == halmos_cheat_code.create_bytes32:
-                    name = name_of(extract_string_argument(arg, 0))
-                    label = f"halmos_{name}_bytes32_{ex.new_symbol_id():>02}"
-                    ret = BitVec(label, BitVecSort256)
-
-                # createAddress(string) returns (address)
-                elif funsig == halmos_cheat_code.create_address:
-                    name = name_of(extract_string_argument(arg, 0))
-                    label = f"halmos_{name}_address_{ex.new_symbol_id():>02}"
-                    ret = uint256(BitVec(label, BitVecSort160))
-
-                # createBool(string) returns (bool)
-                elif funsig == halmos_cheat_code.create_bool:
-                    name = name_of(extract_string_argument(arg, 0))
-                    label = f"halmos_{name}_bool_{ex.new_symbol_id():>02}"
-                    ret = uint256(BitVec(label, BitVecSort1))
-
-                else:
-                    error_msg = f"Unknown halmos cheat code: function selector = 0x{funsig:0>8x}, calldata = {hexify(arg)}"
-                    raise HalmosException(error_msg)
+                ret = halmos_cheat_code.handle(ex, arg)
 
             # vm cheat code
             if eq(to, hevm_cheat_code.address):
                 ex.solver.add(exit_code_var != con(0))
-                # vm.fail()
-                # BitVecVal(hevm_cheat_code.fail_payload, 800)
-                if arg == hevm_cheat_code.fail_payload:
-                    raise FailCheatcode()
-
-                # vm.assume(bool)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg)) == hevm_cheat_code.assume_sig
-                ):
-                    assume_cond = simplify(is_non_zero(Extract(255, 0, arg)))
-                    ex.solver.add(assume_cond)
-                    ex.path.append(str(assume_cond))
-
-                # vm.getCode(string)
-                elif (
-                    simplify(Extract(arg_size * 8 - 1, arg_size * 8 - 32, arg))
-                    == hevm_cheat_code.get_code_sig
-                ):
-                    calldata = bytes.fromhex(hex(arg.as_long())[2:])
-                    path_len = int.from_bytes(calldata[36:68], "big")
-                    path = calldata[68 : 68 + path_len].decode("utf-8")
-
-                    if ":" in path:
-                        [filename, contract_name] = path.split(":")
-                        path = "out/" + filename + "/" + contract_name + ".json"
-
-                    target = self.options["target"].rstrip("/")
-                    path = target + "/" + path
-
-                    with open(path) as f:
-                        artifact = json.loads(f.read())
-
-                    if artifact["bytecode"]["object"]:
-                        bytecode = artifact["bytecode"]["object"].replace("0x", "")
-                    else:
-                        bytecode = artifact["bytecode"].replace("0x", "")
-
-                    ret = stringified_bytes_to_bytes(bytecode)
-                # vm.prank(address)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg)) == hevm_cheat_code.prank_sig
-                ):
-                    result = ex.prank.prank(uint160(Extract(255, 0, arg)))
-                    if not result:
-                        raise HalmosException("You have an active prank already.")
-
-                # vm.startPrank(address)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg))
-                    == hevm_cheat_code.start_prank_sig
-                ):
-                    result = ex.prank.startPrank(uint160(Extract(255, 0, arg)))
-                    if not result:
-                        raise HalmosException("You have an active prank already.")
-
-                # vm.stopPrank()
-                elif (
-                    eq(arg.sort(), BitVecSorts[4 * 8])
-                    and simplify(Extract(31, 0, arg)) == hevm_cheat_code.stop_prank_sig
-                ):
-                    ex.prank.stopPrank()
-                # vm.deal(address,uint256)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32 * 2) * 8])
-                    and simplify(Extract(543, 512, arg)) == hevm_cheat_code.deal_sig
-                ):
-                    who = uint160(Extract(511, 256, arg))
-                    amount = simplify(Extract(255, 0, arg))
-                    ex.balance_update(who, amount)
-                # vm.store(address,bytes32,bytes32)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32 * 3) * 8])
-                    and simplify(Extract(799, 768, arg)) == hevm_cheat_code.store_sig
-                ):
-                    store_account = uint160(Extract(767, 512, arg))
-                    store_slot = simplify(Extract(511, 256, arg))
-                    store_value = simplify(Extract(255, 0, arg))
-                    store_account_addr = self.resolve_address_alias(ex, store_account)
-                    if store_account_addr is not None:
-                        self.sstore(ex, store_account_addr, store_slot, store_value)
-                    else:
-                        raise HalmosException(f"uninitialized account: {store_account}")
-
-                # vm.load(address,bytes32)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32 * 2) * 8])
-                    and simplify(Extract(543, 512, arg)) == hevm_cheat_code.load_sig
-                ):
-                    load_account = uint160(Extract(511, 256, arg))
-                    load_slot = simplify(Extract(255, 0, arg))
-                    load_account_addr = self.resolve_address_alias(ex, load_account)
-                    if load_account_addr is not None:
-                        ret = self.sload(ex, load_account_addr, load_slot)
-                    else:
-                        raise HalmosException(f"uninitialized account: {store_account}")
-
-                # vm.fee(uint256)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg)) == hevm_cheat_code.fee_sig
-                ):
-                    ex.block.basefee = simplify(Extract(255, 0, arg))
-                # vm.chainId(uint256)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg)) == hevm_cheat_code.chainid_sig
-                ):
-                    ex.block.chainid = simplify(Extract(255, 0, arg))
-                # vm.coinbase(address)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg)) == hevm_cheat_code.coinbase_sig
-                ):
-                    ex.block.coinbase = uint160(Extract(255, 0, arg))
-                # vm.difficulty(uint256)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg))
-                    == hevm_cheat_code.difficulty_sig
-                ):
-                    ex.block.difficulty = simplify(Extract(255, 0, arg))
-                # vm.roll(uint256)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg)) == hevm_cheat_code.roll_sig
-                ):
-                    ex.block.number = simplify(Extract(255, 0, arg))
-                # vm.warp(uint256)
-                elif (
-                    eq(arg.sort(), BitVecSorts[(4 + 32) * 8])
-                    and simplify(Extract(287, 256, arg)) == hevm_cheat_code.warp_sig
-                ):
-                    ex.block.timestamp = simplify(Extract(255, 0, arg))
-                # vm.etch(address,bytes)
-                elif extract_funsig(arg) == hevm_cheat_code.etch_sig:
-                    who = extract_bytes(arg, 4 + 12, 20)
-
-                    # who must be concrete
-                    if not is_bv_value(who):
-                        error_msg = f"vm.etch(address who, bytes code) must have concrete argument `who` but received {who}"
-                        raise HalmosException(error_msg)
-
-                    # code must be concrete
-                    try:
-                        code_offset = int_of(extract_bytes(arg, 4 + 32, 32))
-                        code_length = int_of(extract_bytes(arg, 4 + code_offset, 32))
-                        code_int = int_of(
-                            extract_bytes(arg, 4 + code_offset + 32, code_length)
-                        )
-                        code_bytes = code_int.to_bytes(code_length, "big")
-
-                        ex.code[who] = Contract(code_bytes)
-                    except Exception as e:
-                        error_msg = f"vm.etch(address who, bytes code) must have concrete argument `code` but received calldata {arg}"
-                        raise HalmosException(error_msg) from e
-                # ffi(string[]) returns (bytes)
-                elif extract_funsig(arg) == hevm_cheat_code.ffi_sig:
-                    if not self.options.get("ffi"):
-                        error_msg = "ffi cheatcode is disabled. Run again with `--ffi` if you want to enable it"
-                        raise HalmosException(error_msg)
-                    cmd = extract_string_array_argument(arg, 0)
-                    process = Popen(cmd, stdout=PIPE, stderr=PIPE)
-
-                    (stdout, stderr) = process.communicate()
-
-                    if stderr:
-                        warn(
-                            INTERNAL_ERROR,
-                            f"An exception has occurred during the usage of the ffi cheatcode:\n{stderr.decode('utf-8')}",
-                        )
-
-                    out_bytes = stdout.decode("utf-8")
-
-                    if not out_bytes.startswith("0x"):
-                        out_bytes = out_bytes.strip().encode("utf-8").hex()
-                    else:
-                        out_bytes = out_bytes.strip().replace("0x", "")
-
-                    ret = stringified_bytes_to_bytes(out_bytes)
-
-                else:
-                    # TODO: support other cheat codes
-                    msg = f"Unsupported cheat code: calldata = {hexify(arg)}"
-                    raise HalmosException(msg)
+                ret = hevm_cheat_code.handle(self, ex, arg)
 
             # console
             if eq(to, console.address):
                 ex.solver.add(exit_code_var != con(0))
-
-                funsig: int = int_of(
-                    extract_funsig(arg), "symbolic console function selector"
-                )
-
-                if funsig == console.log_uint256:
-                    print(extract_bytes(arg, 4, 32))
-
-                # elif funsig == console.log_string:
-
-                else:
-                    # TODO: support other console functions
-                    print(
-                        color_info(
-                            f"Unsupported console function: selector = 0x{funsig:0>8x}, "
-                            f"calldata = {hexify(arg)}"
-                        )
-                    )
+                console.handle(ex, arg)
 
             # store return value
             if ret_size > 0:
