+++ src/halmos/utils.py
@@ -10,6 +10,108 @@ from z3 import *
 from .exceptions import NotConcreteError
 
 
+Word = Any  # z3 expression (including constants)
+Byte = Any  # z3 expression (including constants)
+Bytes = Any  # z3 expression (including constants)
+Address = BitVecRef  # 160-bitvector
+
+
+# dynamic BitVecSort sizes
+class BitVecSortCache:
+    def __init__(self):
+        self.cache = {}
+        for size in (
+            1,
+            8,
+            16,
+            32,
+            64,
+            128,
+            160,
+            256,
+            264,
+            288,
+            512,
+            544,
+            800,
+            1024,
+            1056,
+        ):
+            self.cache[size] = BitVecSort(size)
+
+    def __getitem__(self, size: int) -> BitVecSort:
+        hit = self.cache.get(size)
+        return hit if hit is not None else BitVecSort(size)
+
+
+BitVecSorts = BitVecSortCache()
+
+# known, fixed BitVecSort sizes
+BitVecSort1 = BitVecSorts[1]
+BitVecSort8 = BitVecSorts[8]
+BitVecSort160 = BitVecSorts[160]
+BitVecSort256 = BitVecSorts[256]
+BitVecSort264 = BitVecSorts[264]
+BitVecSort512 = BitVecSorts[512]
+
+
+def concat(args):
+    if len(args) > 1:
+        return Concat(args)
+    else:
+        return args[0]
+
+
+def uint256(x: BitVecRef) -> BitVecRef:
+    bitsize = x.size()
+    if bitsize > 256:
+        raise ValueError(x)
+    if bitsize == 256:
+        return x
+    return simplify(ZeroExt(256 - bitsize, x))
+
+
+def uint160(x: BitVecRef) -> BitVecRef:
+    bitsize = x.size()
+    if bitsize > 256:
+        raise ValueError(x)
+    if bitsize == 160:
+        return x
+    if bitsize > 160:
+        return simplify(Extract(159, 0, x))
+    else:
+        return simplify(ZeroExt(160 - bitsize, x))
+
+
+def con(n: int, size_bits=256) -> Word:
+    return BitVecVal(n, BitVecSorts[size_bits])
+
+
+#             x  == b   if sort(x) = bool
+# int_to_bool(x) == b   if sort(x) = int
+def test(x: Word, b: bool) -> Word:
+    if is_bool(x):
+        if b:
+            return x
+        else:
+            return Not(x)
+    elif is_bv(x):
+        if b:
+            return x != con(0)
+        else:
+            return x == con(0)
+    else:
+        raise ValueError(x)
+
+
+def is_non_zero(x: Word) -> Word:
+    return test(x, True)
+
+
+def is_zero(x: Word) -> Word:
+    return test(x, False)
+
+
 def create_solver(logic="QF_AUFBV", ctx=None, timeout=0, max_memory=0):
     # QF_AUFBV: quantifier-free bitvector + array theory: https://smtlib.cs.uiowa.edu/logics.shtml
     solver = SolverFor(logic, ctx=ctx)
@@ -22,6 +124,33 @@ def create_solver(logic="QF_AUFBV", ctx=None, timeout=0, max_memory=0):
         solver.set(max_memory=max_memory)
 
     return solver
+
+
+def extract_bytes(data: BitVecRef, byte_offset: int, size_bytes: int) -> BitVecRef:
+    """Extract bytes from calldata. Zero-pad if out of bounds."""
+    n = data.size()
+    if n % 8 != 0:
+        raise ValueError(n)
+
+    # will extract hi - lo + 1 bits
+    hi = n - 1 - byte_offset * 8
+    lo = n - byte_offset * 8 - size_bytes * 8
+    lo = 0 if lo < 0 else lo
+
+    val = simplify(Extract(hi, lo, data))
+
+    zero_padding = size_bytes * 8 - val.size()
+    if zero_padding < 0:
+        raise ValueError(val)
+    if zero_padding > 0:
+        val = simplify(Concat(val, con(0, zero_padding)))
+
+    return val
+
+
+def extract_funsig(calldata: BitVecRef):
+    """Extracts the function signature (first 4 bytes) from calldata"""
+    return extract_bytes(calldata, 0, 4)
 
 
 def bv_value_to_bytes(x: BitVecNumRef) -> bytes:
@@ -113,7 +242,7 @@ def stringify(symbol_name: str, val: Any):
         return f"0x{val.as_long():0{byte_length(val) * 2}x}"
     elif type_name == "bool":
         return str(val.as_long() != 0).lower()
-    elif type_name == "string":
+    elif type_name == "string" or "string" in symbol_name:  # XXX TEMP HACK
         str_val = bytes.fromhex(hexify(val)[2:]).decode("utf-8")
         return f'"{str_val}"'
     elif type_name == "bytes":
